requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
executing==2.2.1
fastapi==0.131.0
fastjsonschema==2.21.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
langchain==1.2.10
langchain-core==1.2.14
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.8
langsmith==0.7.6
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.2.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
openai==2.21.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.9.2
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
stack-data==0.6.3
starlette==0.52.1
tenacity==9.1.4
tinycss2==1.4.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
wcwidth==0.6.0
webencodings==0.5.1
xxhash==3.6.0
yarg==0.1.9
zstandard==0.25.0
